+++ lib/infinity_one_web/controllers/site_avatar_controller.ex
@@ -5,7 +5,6 @@ defmodule InfinityOneWeb.SiteAvatarController do
   """
   use InfinityOneWeb, :controller
 
-  alias InfinityOne.Accounts
   alias InfinityOne.Settings.General
   alias InfinityOne.SiteAvatar