+++ plugins/one_chat/priv/repo/migrations/04420180112043010_add_patters_to_chat_general.exs
@@ -1,0 +1,9 @@
+defmodule InfinityOne.Repo.Migrations.AddPatternsToChatGeneral do
+  use Ecto.Migration
+
+  def change do
+    alter table(:settings_chat_general) do
+      add :message_replacement_patterns, :text
+    end
+  end
+end